# postman/Student Management.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Student Management",
    "description": "Complete Spring Boot Student Management System API collection.\n\n1. Run Login (or Register) first.\n2. Token is saved automatically from the login/register response.\n3. Create Classroom → Subject → Student → Score in that order.\n4. Use returned IDs in path variables (or update collection variables).",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get User",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": [
                  {
                    "email": "admin@school.com",
                    "id": 1,
                    "role": "ADMIN",
                    "username": "admin"
                  },
                  {
                    "email": "teacher1@school.com",
                    "id": 2,
                    "role": "TEACHER",
                    "username": "teacher"
                  }
                ]
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "Add User",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "username": "admin1",
                "password": "1admin123",
                "email": "admin1@school.com",
                "role": "ADMIN"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": "User created successfully"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/classrooms": {
      "post": {
        "summary": "Create Classroom",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "classroomCode": {
                    "type": "string"
                  },
                  "classroomName": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  }
                },
                "required": [
                  "classroomCode",
                  "classroomName",
                  "capacity"
                ]
              },
              "example": {
                "classroomCode": "C001",
                "classroomName": "Grade 10A",
                "capacity": 40
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": []
      },
      "get": {
        "summary": "Get Classrooms",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "example": "0",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": "10",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "",
            "required": true,
            "example": "id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "description": "",
            "required": true,
            "example": "asc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/classrooms/{id}/students": {
      "get": {
        "summary": "Get Student By ClassroomID",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "content": [
                    {
                      "id": 2,
                      "studentId": "STU002",
                      "fullName": "Rack Klin",
                      "dateOfBirth": "2002-02-20",
                      "gender": "FEMALE",
                      "email": "rack.klin@example.com",
                      "address": "123 Main Street, Mandalay",
                      "parentPhone": "09123456789",
                      "enrollmentDate": "2026-07-17",
                      "status": "ACTIVE",
                      "classroomId": 1,
                      "classroomCode": null,
                      "classroomName": "Grade 10 A"
                    }
                  ],
                  "empty": false,
                  "first": true,
                  "last": true,
                  "number": 0,
                  "numberOfElements": 1,
                  "pageable": {
                    "offset": 0,
                    "pageNumber": 0,
                    "pageSize": 10,
                    "paged": true,
                    "sort": {
                      "empty": false,
                      "sorted": true,
                      "unsorted": false
                    },
                    "unpaged": false
                  },
                  "size": 10,
                  "sort": {
                    "empty": false,
                    "sorted": true,
                    "unsorted": false
                  },
                  "totalElements": 1,
                  "totalPages": 1
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "Assign Students by ClassroomID",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "studentIds": [
                  1
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": "Students assigned successfully"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/classrooms/{id}/students/{studentId}": {
      "delete": {
        "summary": "Remove Students From Classroom",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": "Student removed from classroom successfully"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}